+++ src/pages/DataPage.svelte
@@ -7,7 +7,6 @@
 <script lang="ts">
   import Gesture from '../components/Gesture.svelte';
   import { state } from '../script/stores/uiStore';
-  import { downloadDataset, loadDatasetFromFile } from '../script/stores/mlStore';
   import { t } from '../i18n';
   import RecordInformationContent from '../components/datacollection/RecordInformationContent.svelte';
   import StandardDialog from '../components/dialogs/StandardDialog.svelte';
@@ -20,6 +19,8 @@
   import Information from '../components/information/Information.svelte';
   import { onMount } from 'svelte';
   import { gestures } from '../script/stores/Stores';
+    import FileUtility from '../script/repository/FileUtility';
+    import { get } from 'svelte/store';
 
   let isConnectionDialogOpen = false;
 
@@ -37,7 +38,7 @@
   };
 
   const onDownloadGestures = () => {
-    downloadDataset();
+    FileUtility.downloadDataset(get(gestures));
   };
 
   const onUploadGestures = () => {
@@ -46,6 +47,7 @@
 
   let filePicker: HTMLInputElement;
   onMount(() => {
+    // Todo: Maybe move some of this to the file utility class
     filePicker = document.createElement('input');
     filePicker.type = 'file';
     filePicker.accept = 'application/JSON';
@@ -54,7 +56,7 @@
         return;
       }
       const f = filePicker.files[0];
-      loadDatasetFromFile(f);
+      FileUtility.loadDatasetFromFile(f);
       filePicker.value = ''; // To trick element to trigger onChange if same file selected
     };
     return () => {
