+++ src/script/repository/FileUtility.ts
@@ -1,0 +1,38 @@
+import { GestureData } from "../domain/Gesture";
+import { PersistantGestureData } from "../domain/Gestures";
+import { gestures } from "../stores/Stores";
+
+class FileUtility {
+    public static loadDatasetFromFile(file: File) {
+    const reader = new FileReader();
+    reader.onload = function (e: ProgressEvent<FileReader>) {
+      if (!e.target) {
+        return;
+      }
+      const contents = e.target.result;
+      if (typeof contents === 'string') {
+        gestures.importFrom(JSON.parse(contents) as PersistantGestureData[]);
+      }
+    };
+    reader.readAsText(file as Blob);
+  }
+  
+  public static downloadDataset(gestureData: GestureData[]) {
+    const element = document.createElement('a');
+    element.setAttribute(
+      'href',
+      'data:application/json;charset=utf-8,' +
+        encodeURIComponent(JSON.stringify(gestureData, null, 2)),
+    );
+    element.setAttribute('download', 'dataset');
+  
+    element.style.display = 'none';
+    document.body.appendChild(element);
+  
+    element.click();
+    document.body.removeChild(element);
+  }
+  
+}
+
+export default FileUtility;
